Changed NodeId underlying type to be uint16_t for now

+++ src/GraphTypes.h
@@ -8,7 +8,7 @@
 namespace graphski
 {
 	// global alias for node id type
-	using NodeId = uint8_t;
+	using NodeId = uint16_t;
 	// pair of node id and edges id inside node's edges vector
 	using EdgeId = std::pair<NodeId, NodeId>;
